fix: various improvements & bug fixes using Claude

+++ entities/item.go
@@ -38,7 +38,7 @@ type DroppedItem struct {
 	EntityId    int32
 	ItemId      int32
 	Amount      byte
-	Metadata    byte
+	Metadata    uint16
 	X, Y, Z     float64
 	PickupDelay int32
 	Dim         int32
